Mass - (r)mag relation
Motivation: To sample many lum functions, it is necessary to compute
dN/dlogM quickly for different sets of parameters so best not to
rely on pre-computed luminosity functions from isochrones.
To calculate luminosity functions on the fly, the first step is to
pick an isochrone and find the relation between mass and magnitude,
as the theoretical LFs are given in terms of mass, not magnitude,
while the fitting is done in dN/dmag space.

- Created test_lf.py: program takes an isochrone, finds a spline
  between mass and magnitude (r), and calculates a spline in
  mass - magnitude space (currently x-axis = mass). The spline
  can also be used to calculate the derivative (see dr_new).

- added mass column to read_iso(), will also update it on main.py
  later.