sim...cmd.py: Simulated CMD(no observ. errors yet)
Continued development on simulate_cmd.py

*While developing this code identified issue with read_iso_darth,
see previous commit.

*mass cuts: Eliminated code snippet that reset mass_min and mass_max
in the case that the requested values were outside of the bounds of
the isochrone. this was superflous. *if* the mass limits are outside
the bounds of the isochrone, then when creating dummy samples in M,
dn/dM space (Monte Carlo samples), then I throw away those samples
where M is outside of the bounds of the isochrone.

*interpolated isochrone mass-magnitude relation:
To map a simulated datum's mass to its magnitude, need to find a
relation between mass and magnitude for the isochrone in use. The
code fits a spline to the (mass-sorted) mass - magnitude columns
of the isochrone data frame. The spline is an interpolating spline
passing through *all* data points - other spline options include
using only one every n data points.

*For each simulated star, simulate_cmd.py currently calculates
two magnitudes, and a color (mag1 - mag2). In the case of the
wfpc2 data, mag1=f606w, and mag2=f814w.

*diagnostic plots can be seen by setting testing = 1 at the
beginning of the module simulate_cmd(...)