f_salpeter: change of min_mass and mass_max
The dN/dM function is no longer truncated for M
outside of M_min and M_max, only outside of 0.1Msun
and 100 Msun.

M_min and M_max are instead *normalized* using
only the integral of dN/dM * deltaM between
M_min and M_max. The normalization is then
applied to the entire range of masses in the
isochrone.

The rationale is that cuts should only be
done on the data, and not on the isochrone.